+++ python_packages/ballsbot_localization/src/free_distances.h
@@ -4,4 +4,4 @@
 using FreeDistances = DirectionsWeights;
 
 FreeDistances GetFreeDistances(const PointCloud& points, CarInfo car_info);
-FreeDistances DebugGetFreeDistances(const Grid& grid, int current_ts, CarInfo car_info);
+FreeDistances DebugGetFreeDistances(const Grid& grid, double current_ts, CarInfo car_info);
